+++ Model/Path/pathxmlparser.cpp
@@ -113,14 +113,14 @@ QString PathXMLParser::readNameElement(QXmlStreamReader& xmlReader){
     return nameElement;
 }
 
-float PathXMLParser::readCoordinateElement(QXmlStreamReader &xmlReader){
-    float coordinate(0.0);
+double PathXMLParser::readCoordinateElement(QXmlStreamReader &xmlReader){
+    double coordinate(0.0);
     while(!xmlReader.atEnd()){
         if(xmlReader.isEndElement()){
             xmlReader.readNext();
             break;
         } else if(xmlReader.isStartElement()){
-            coordinate = xmlReader.readElementText().toFloat();
+            coordinate = xmlReader.readElementText().toDouble();
             xmlReader.readNext();
             break;
         } else if(xmlReader.isCharacters())
