+++ View/Robot/obstaclespainteditem.h
@@ -7,9 +7,9 @@
 class ObstaclesPaintedItem : public QQuickPaintedItem {
 
     Q_OBJECT
-    Q_PROPERTY(float orientation_ READ orientation WRITE setOrientation)
-    Q_PROPERTY(float _x READ getX WRITE setX)
-    Q_PROPERTY(float _y READ getY WRITE setY)
+    Q_PROPERTY(double orientation_ READ orientation WRITE setOrientation)
+    Q_PROPERTY(double _x READ getX WRITE setX)
+    Q_PROPERTY(double _y READ getY WRITE setY)
 
 public:
 
@@ -26,7 +26,7 @@ public:
      * @param ranges
      * To update the display of the obstacles on the map when we receive it from the robot
      */
-    void updateObstacles(float angle_min, float angle_max, float angle_increment, QVector<float> ranges);
+    void updateObstacles(double angle_min, double angle_max, double angle_increment, QVector<double> ranges);
 
     /**
      * @brief clearObstacles
@@ -34,19 +34,19 @@ public:
      */
     void clearObstacles(bool activated);
 
-    float orientation() const { return orientation_; }
-    float getX() const { return _x; }
-    float getY() const { return _y; }
+    double orientation() const { return orientation_; }
+    double getX() const { return _x; }
+    double getY() const { return _y; }
 
-    void setOrientation(const float orientation) { orientation_ = orientation; }
-    void setX(const float x) { _x = x; }
-    void setY(const float y) { _y = y; }
+    void setOrientation(const double orientation) { orientation_ = orientation; }
+    void setX(const double x) { _x = x; }
+    void setY(const double y) { _y = y; }
 
 private:
     QVector<QPointF> obstacles_;
-    float orientation_;
-    float _x;
-    float _y;
+    double orientation_;
+    double _x;
+    double _y;
     bool activated;
 };
