+++ View/Robot/scanmappainteditem.h
@@ -10,9 +10,9 @@ class ScanMapPaintedItem : public QQuickPaintedItem {
 
         Q_OBJECT
 
-    Q_PROPERTY(float orientationRobot READ robotOrientation WRITE setRobotOrientation NOTIFY orientationRobotChanged)
-    Q_PROPERTY(float xRobot READ robotX WRITE setRobotX NOTIFY xRobotChanged)
-    Q_PROPERTY(float yRobot READ robotY WRITE setRobotY NOTIFY yRobotChanged)
+    Q_PROPERTY(double orientationRobot READ robotOrientation WRITE setRobotOrientation NOTIFY orientationRobotChanged)
+    Q_PROPERTY(double xRobot READ robotX WRITE setRobotX NOTIFY xRobotChanged)
+    Q_PROPERTY(double yRobot READ robotY WRITE setRobotY NOTIFY yRobotChanged)
     Q_PROPERTY(QString ip READ getIp WRITE setIp NOTIFY ipChanged)
     Q_PROPERTY(bool _drawRobotView READ drawRobotView WRITE setDrawRobotView NOTIFY drawRobotViewChanged)
 
@@ -30,18 +30,18 @@ public:
 
     QImage& getImage(void) { return _image; }
 
-    float robotOrientation(void) const { return orientationRobot; }
-    float robotX(void) const { return xRobot; }
-    float robotY(void) const { return yRobot; }
+    double robotOrientation(void) const { return orientationRobot; }
+    double robotX(void) const { return xRobot; }
+    double robotY(void) const { return yRobot; }
     QString getIp(void) const { return ip; }
     bool drawRobotView(void) const { return _drawRobotView; }
     int getLeft(void) const { return left; }
     int getTop(void) const { return top; }
 
-    void setRobotX(const float x);
-    void setRobotY(const float y);
+    void setRobotX(const double x);
+    void setRobotY(const double y);
 
-    void setRobotOrientation(const float ori) { orientationRobot = ori; emit updateRobot(); }
+    void setRobotOrientation(const double ori) { orientationRobot = ori; emit updateRobot(); }
     void setIp(const QString _ip) { ip = _ip; }
     void setDrawRobotView(const bool draw) {
                                             _drawRobotView = draw;
@@ -64,9 +64,9 @@ private:
     QString ip;
     int left;
     int top;
-    float xRobot;
-    float yRobot;
-    float orientationRobot;
+    double xRobot;
+    double yRobot;
+    double orientationRobot;
     QImage _image;
 };
